# seems to make the latest code faster by reducing the back-n-forth between JNI and Java.