re: https://github.com/xerial/sqlite-jdbc/issues/160

On my MacBook Pro,
**3.8.11.2**
> Time to create & insert: 6872 ms
> Time to select: 8963 ms
> Time to select & get resultset values: 10279 ms

**3.14.2**
> Time to create & insert: 6654 ms
> Time to select: 11362 ms
> Time to select & get resultset values: 14068 ms
> 

**3.14.2.1-SNAPSHOT**
> Time to create & insert: 6527 ms
> Time to select: 8942 ms
> Time to select & get resultset values: 10936 ms
>