@xerial 2016-09-30
Merged. Thanks!